+++ CHANGELOG.md
@@ -17,6 +17,7 @@ with `zswap` being tracked in [Changelog Zswap](./CHANGELOG_zswap.md).
 - fix: fix potential panic in MPT path removal, unlikely to be currently triggerable.
 - fix: fix potential panic in bridge fee processing
 - fix: address non-associativity of Dust event processing.
+- fix: tighten cost heuristic, no longer being overly conservative, moving less transactions to the fallible section.
 
 ## 8.1.0